refactor excalidraw
