+++ docusaurus.config.ts
@@ -15,13 +15,13 @@ import {
   rsDoctorPluginConfig,
   aliasConfigurationPluginConfig,
   sentryPluginConfig,
-  excalidrawPluginConfig,
   socketIoNoDepWarningsPluginConfig,
 } from './src/siteConfig/pluginConfigs';
 import { useTdevContentPath } from './src/siteConfig/helpers';
 import path from 'path';
 import { recommendedBeforeDefaultRemarkPlugins, recommendedRehypePlugins, recommendedRemarkPlugins } from './src/siteConfig/markdownPluginConfigs';
 import { remarkPdfPluginConfig } from '@tdev/remark-pdf';
+import { excalidrawPluginConfig } from '@tdev/excalidoc';
 
 const siteConfig = getSiteConfig();
