+++ packages/excalidoc/Component/Preview/Image.tsx
