+++ packages/excalidoc/Component/Preview/images/excalidraw-logo.png
