+++ packages/excalidoc/Component/styles.module.scss
