+++ tsconfig.json
@@ -4,7 +4,7 @@
     "compilerOptions": {
         "baseUrl": ".",
         "experimentalDecorators": false,
-        "types": ["docusaurus-plugin-sass", "@docusaurus/module-type-aliases"],
+        "types": ["docusaurus-plugin-sass", "@docusaurus/module-type-aliases", "@types/node"],
         "module": "esnext",
         "target": "esnext",
         "strict": true,
